Change namespace to markov::tram. Fix the parallel fors. Refactor.
